Configure signIn method
